make prettier override some eslint rules
putting plugins at the bottom of the config makes them apply last, which is what we want. and putting prettier at the bottom of the extends array makes the prettier rules apply last, which is also what we want. sourced from this stack overflow answer (with useful replies) https://stackoverflow.com/a/56350664