build: move project version to v2
This adds v2 directories where necessary and all old code moved to v1 directories where it wasn't
already. Updated imports to v1 files and tested before setting prisma to use v2 schema.

BREAKING CHANGE: This breaks the state of the application because it's expecting a Prisma schema to
be in the v2 directory, of which there are none currently. This is a non-issue since the application
has not been deployed yet.

fixes #3