refactor: change UserSchema to a curried function
zod has no functionallity to refer to an adjacent value in the input object to compare two properties, so comparing password and compare fields is not trivial without creating a regex on the fly, same goes for checking that the email address starts with the username. This function creates regex when called with the input object so that this is possible. Also added custom error messages where necessary.